fix(update): harden transactional rollback and locking
